Handle hitsounds in the screen rather than the ruleset
